chore(textSize): Refactor textSize out of Widget
